organize and add tests
